requirements.txt
numpy
pandas
scikit-learn
scipy
xgboost
imblearn
joblib
python-dateutil
pytz
threadpoolctl
tzdata
six
python-dotenv
requests
pdfplumber
matplotlib
seaborn
plotly
folium
streamlit
streamlit-folium
Flask
click
colorama
Werkzeug
itsdangerous
Jinja2
MarkupSafe
blinker
langchain
langchain-community
huggingface-hub
ibm-watson
ibmcloudant
firebase-admin
selenium
webdriver-manager
openai
python-dotenv